+++ client/components/projects/Project.jsx
@@ -1,6 +1,5 @@
 // @flow
 import React from 'react';
-import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { getMotivationData } from '../../actions/dashboard';
 import { clearLogs } from '../../actions/timelog';
@@ -10,11 +9,8 @@ import Layout from '../layout/Layout';
 import Spinner from '../layout/Spinner';
 import NotFound from '../NotFound';
 import TasksList from '../tasks/TasksList';
-import AddForm from './AddForm';
-import MotivationBlock from './MotivationBlock';
-import TimeAddForm from '../tasks/TimeAddForm';
-import { formatTime } from '../../helpers';
 import { IMatch } from '../../types';
+import ProjectInfo from './ProjectInfo';
 
 type ProjectProps = {
   match: IMatch,
@@ -28,23 +24,12 @@ type ProjectProps = {
   clearSecondsLog: () => void
 };
 
-type State = {
-  showForm: boolean,
-  timeForm: boolean,
-  newTaskInput: string
-};
-
-class Project extends React.Component<ProjectProps, State> {
+class Project extends React.Component<ProjectProps> {
   static defaultProps = {
     projects: [],
     tasks: [],
     user: {},
     dashboardData: {}
-  };
-  state = {
-    showForm: false,
-    timeForm: false,
-    newTaskInput: ''
   };
 
   componentDidMount() {
@@ -56,38 +41,6 @@ class Project extends React.Component<ProjectProps, State> {
   componentWillUnmount() {
     this.props.clearSecondsLog();
   }
-
-  showAddForm = () => {
-    this.setState({
-      showForm: !this.state.showForm
-    });
-  };
-
-  showAddTimeForm = () => {
-    this.setState({
-      timeForm: !this.state.timeForm
-    });
-  };
-
-  handleNewTaskInput = (e: SyntheticEvent<HTMLInputElement>) => {
-    this.setState({
-      newTaskInput: e.currentTarget.value
-    });
-  };
-
-  addTask = task => {
-    this.props.handleNewTask({ name: task.name, project: task.project });
-    this.setState({
-      showForm: false,
-      newTaskInput: ''
-    });
-  };
-
-  formToggle = () => {
-    this.setState(state => ({
-      timeForm: !state.timeForm
-    }));
-  };
 
   render() {
     const { dashboardData, tasks } = this.props;
@@ -101,75 +54,16 @@ class Project extends React.Component<ProjectProps, State> {
     if (tasks.project === false) {
       return <NotFound />;
     }
-    const { showForm, newTaskInput } = this.state;
-    const addLinkText = showForm ? 'Hide The Form' : 'New Task';
-    const projectTime = tasks.project.timeSpent;
-
-    const title = tasks.project.name;
 
     return (
       <Layout>
-        <div className="project__description">
-          <div className="project__info">
-            <h1 className="page-title">{title}</h1>
-            <h3 className="page-title">
-              <span className="pretty-time">
-                Total: {formatTime(projectTime)}
-              </span>
-            </h3>
-            <div className="project__buttons">
-              <button onClick={this.showAddForm} className="button--submit">
-                {addLinkText}
-              </button>
-              <button
-                onClick={this.showAddTimeForm}
-                className="button--submit"
-                title="Add Time To The Project"
-              >
-                Add Time
-              </button>
-            </div>
-
-            {this.state.timeForm ? (
-              <TimeAddForm
-                addProjectTime={this.props.handleAddingTimeToProject}
-                from="project"
-                formToggle={this.formToggle}
-                project={tasks.project._id}
-              />
-            ) : null}
-
-            {this.state.showForm ? (
-              <AddForm
-                inputValue={newTaskInput}
-                className="form form__newtask"
-                handleInput={this.handleNewTaskInput}
-                clickHandler={e => {
-                  e.preventDefault();
-                  this.addTask({
-                    name: newTaskInput,
-                    project: this.props.match.params.id
-                  });
-                }}
-                labelName="Name"
-              />
-            ) : null}
-          </div>
-          <div className="project__motivation">
-            <div className="motivation-paragraph">
-              <p>
-                You can checkout tasks you already done{' '}
-                <Link to={`${this.props.location.pathname}/archive`}>here</Link>
-                .
-              </p>
-              <MotivationBlock
-                dashboardData={dashboardData}
-                initialTime={tasks.project.initialTime}
-                time={tasks.project.timeSpent}
-              />
-            </div>
-          </div>
-        </div>
+        <ProjectInfo
+          project={tasks.project}
+          handleAddingTimeToProject={this.props.handleAddingTimeToProject}
+          handleNewTask={this.props.handleNewTask}
+          pathname={this.props.location.pathname}
+          dashboardData={dashboardData}
+        />
         <TasksList projectId={this.props.match.params.id} />
       </Layout>
     );
